requirements.txt
MetaTrader5
pandas
numpy
python-dotenv
pypdf
psutil
